% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp500}
\alias{sp500}
\title{504 SP500 stocks}
\format{
A data frame with 504 rows and 3 variables:
\describe{
\item{Ticker}{Ticker}
\item{Company}{Company Name}
\item{Industry}{Industry}
\item{Updated_Date}{Date}
}
}
\source{
\url{https://www.liberatedstocktrader.com/sp-500-companies-list-alphabetically-sorted/}
}
\usage{
sp500
}
\description{
A dataset containing ticker, company name, and industry of 504 stocks in SP500
}
\keyword{datasets}
