% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wheel_extractor.R
\name{wheel_extractor}
\alias{wheel_extractor}
\title{Bind two factors}
\usage{
wheel_extractor(
  ticker = c("AAPL", "GOOG"),
  expiration_date = Sys.Date(),
  consumer_key = NULL,
  sleep_second = 1
)
}
\arguments{
\item{ticker}{character}

\item{expiration_date}{date}

\item{consumer_key}{character}

\item{sleep_second}{integer}
}
\value{
factor
}
\description{
Create a new factor from two existing factors, where the new factor's levels
are the union of the levels of the input factors.
}
