% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wheel_visualizer.R
\name{wheel_visualizer}
\alias{wheel_visualizer}
\title{Find the Best Options for The Wheel Strategy}
\usage{
wheel_visualizer(
  data = NULL,
  visualization_type = "ggplot_histogram",
  include_in_the_money = FALSE
)
}
\arguments{
\item{data}{xx}

\item{visualization_type}{yy}

\item{include_in_the_money}{zz}
}
\value{
A data frame object with 10 columns: bid
}
\description{
Create a data frame with options information to determine
what options is the best for short put also known as "The Wheel" strategy
}
