% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wheel_visualizer.R
\name{wheel_visualizer}
\alias{wheel_visualizer}
\title{Visualize Options with the Best Return on Risk}
\usage{
wheel_visualizer(
  data = NULL,
  visualization_type = "ggplot_histogram",
  include_in_the_money = FALSE
)
}
\arguments{
\item{data}{A resulting data frame object from wheel_extractor() function.}

\item{visualization_type}{Currently only support ggplot2's ggplot_histogram.}

\item{include_in_the_money}{The Wheel generally employs on OTM options. But the data frame also includes ITM.}
}
\value{
Ggplot2 histogram chart.
}
\description{
This function will create a histogram plot for the resulting data frame from
the wheel_extractor() function.
}
\examples{
#First let's create sample data frame
sample <- data.frame(
     strikePrice = c(50, 60, 70, 80),
     inTheMoney = c(FALSE, FALSE, FALSE, TRUE),
     ticker = c('AAPL', 'AAPL', 'TSLA', 'TSLA'),
     prob_otm = c(0.3, 0.4, 0.3, 0.4),
     cash_inflow = c(250, 400, 115, 220),
     invested_capital = c(5000, 6000, 7000, 8000),
     return_on_risk = c(0.05, 0.07, 0.02, 0.03)
)

#Then use the function
wheel_visualizer(data = sample)

}
